서버코드: N/A 클라이언트: 채팅 말풍선 방향 조정 완료

+++ app/src/main/java/com/the/war/of/thewarofstars/model/response/ChattingResponse.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class ChattingResponse(
 
     @SerializedName("commentDate")
-    val commentDate: Int
+    val commentDate: Long
 )
